Use temporary workspace for QuickOpenTest
* `Explorer.LocateBufferTest` changes the workspace to a
temporary workspace, but `QuickOpenTest` wanted the Oni source directory.
* So use a temporary workspace for `QuickOpenTest` as well and commonise
the construction of such workspaces into `Common.ts`.
* Also removed some debug logging.
