+++ test/ci/QuickOpenTest.ts
@@ -1,9 +1,11 @@
+import * as fs from "fs"
+
 import * as Oni from "oni-api"
 
 import * as path from "path"
 
 import { Assertor } from "./Assert"
-import { waitForCommand } from "./Common"
+import { useTempWorkspace, waitForCommand } from "./Common"
 
 export const test = async (oni: Oni.Plugin.Api) => {
     const assert = new Assertor("QuickOpen", oni)
@@ -11,6 +13,8 @@ export const test = async (oni: Oni.Plugin.Api) => {
     await oni.automation.waitFor(() => oni.plugins.loaded)
     const instance = oni.plugins.getPlugin("oni-plugin-quickopen")
     assert.defined(instance, "plugin instance")
+
+    await useTempWorkspace(oni)
 
     // Start
     console.info(`QuickOpen: Open and close the menu`) // tslint:disable-line no-console
@@ -31,13 +35,10 @@ export const test = async (oni: Oni.Plugin.Api) => {
 
     const anyOni = oni as any
     await oni.automation.sleep(1000)
-    anyOni.automation.sendKeysV2("searchprovider.ts")
+    anyOni.automation.sendKeysV2("file.ts")
     await oni.automation.sleep(2000)
     anyOni.automation.sendKeysV2("<CR>")
     const editor = oni.editors.activeEditor
-    await assert.waitFor(
-        () => editor.activeBuffer.filePath,
-        (fp: string) => fp.endsWith("SearchProvider.ts"),
-    )
+    await assert.waitFor(() => editor.activeBuffer.filePath, (fp: string) => fp.endsWith("File.ts"))
     // End
 }
